Adds dropdown for level difficulty
I added a dropdown to options menu for the three levels of difficulty: EASY, MEDIUM, HARD. I added some functionality to switch between main menu and options menu. I tried to figure out a way to save the selected level so it carries into the main scene. I decided to write to a file called Level.txt and I added a method in WordListImport that reads from that file, and based on what number it reads it will choose a different word list csv file. Right now if you choose level MEDIUM or HARD it breaks because we do not currently have csv files for these yet.